# docs/endpoints/openapi-systems-ops.yaml
openapi: 3.0.3
info:
  title: Atlas API — Systems Ops
  version: '1.0.0'
  description: |
    Contrato OpenAPI ULTRA DETALLADO, generado leyendo directamente el código fuente
    (controladores, schemas Zod, mappers, servicios y repositorios), del módulo
    `systems-ops` del backend Atlas: catálogo técnico de endpoints/herramientas/entidades
    de datos, log de acciones (auditoría HTTP), cola de revisión de gobierno de datos,
    perfiles y runs de stress testing, y suites/pasos/runs de pruebas E2E internas.

    ## Envolturas globales (aplican a TODOS los endpoints de este documento)

    Este backend registra dos providers globales en `src/app.module.ts`
    (`APP_INTERCEPTOR` y `APP_FILTER`) que envuelven **toda** respuesta, éxito o error.
    Los schemas de `data` documentados endpoint por endpoint abajo son el contenido
    de la propiedad `data`, **no** el body HTTP completo.

    **Éxito** (`ResponseInterceptor`, `src/common/interceptors/response.interceptor.ts`):
    ```json
    { "requestId": "corr-id-o-undefined", "data": { "...": "..." }, "timestamp": "2026-07-07T12:00:00.000Z" }
    ```

    **Error** (`HttpExceptionFilter`, `src/common/filters/http-exception.filter.ts`):
    ```json
    { "requestId": "corr-id-o-undefined", "error": { "code": "VALIDATION_ERROR", "message": "..." }, "timestamp": "2026-07-07T12:00:00.000Z" }
    ```
    `error.code` se deriva del status HTTP: `400→VALIDATION_ERROR`, `401→UNAUTHORIZED`,
...
    de query/params/body.

...

...
    ```ts
...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /systems/action-logs:
    get:
      operationId: listActionLogs
      tags: [SystemsActionLog]
      summary: Listar log de auditoría HTTP con filtros y paginación
      parameters:
        - $ref: '#/components/parameters/ActionLogEndpointIdQuery'
        - $ref: '#/components/parameters/ActionLogRequestIdQuery'
        - $ref: '#/components/parameters/ActionLogCorrelationIdQuery'
        - $ref: '#/components/parameters/ActionLogMethodQuery'
        - $ref: '#/components/parameters/ActionLogStatusCodeQuery'
        - $ref: '#/components/parameters/ActionLogActorTypeQuery'
        - name: module
          in: query
          schema: { type: string, minLength: 1, maxLength: 120 }
          description: Filtra por módulo del endpoint asociado.
        - $ref: '#/components/parameters/RiskLevelQuery'
        - name: containsPii
          in: query
          schema: { type: boolean }
          description: Filtra logs que contienen PII (coercionado desde string).
        - name: from
          in: query
          schema: { type: string, format: date-time }
          description: Rango de fecha desde (ISO 8601).
        - name: to
          in: query
          schema: { type: string, format: date-time }
          description: Rango de fecha hasta (ISO 8601).
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Lista paginada de action logs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ActionLogList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /systems/action-logs/request/{requestId}:
    get:
      operationId: getActionLogsByRequestAlias
      tags: [SystemsActionLog]
      summary: 'Alias de compatibilidad: logs asociados a un requestId'
      description: >-
        Handler idéntico a `GET /systems/action-logs/by-request/{requestId}` (mismo método de
        servicio `getActionLogsByRequest`). Si no existe ningún log con ese requestId, NO lanza
        404 — devuelve `items` como array vacío.
      parameters:
        - $ref: '#/components/parameters/RequestIdPath'
      responses:
        '200':
          description: Logs asociados al requestId (orden occurredAt DESC), sin envelope de paginación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ActionLogItemsOnly'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /systems/action-logs/by-request/{requestId}:
    get:
      operationId: getActionLogsByRequest
      tags: [SystemsActionLog]
      summary: Ruta canónica (no-alias) — logs asociados a un requestId
      parameters:
        - $ref: '#/components/parameters/RequestIdPath'
      responses:
        '200':
          description: Logs asociados al requestId, misma forma que el alias.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ActionLogItemsOnly'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ============================================================
  # SystemsCatalogController
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT emitido por `POST /api/v1/internal/auth/login` (o `/api/v1/auth/login` para
        roles no internos). Debe pertenecer a un usuario con rol en:
        system_admin, platform_admin, admin, qa_engineer, devops, risk_analyst,
        compliance_analyst, readonly_auditor.
  responses:
    BadRequest:
      description: Entrada inválida (Zod) o regla de negocio violada — ver detalle por endpoint.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Falta `Authorization`, formato inválido, token expirado/inválido o revocado.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: 'Rol autenticado no está en la lista permitida: "El usuario autenticado no tiene permiso para esta operación."'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  parameters:
    PageQuery:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: Número de página (entero positivo, coercionado desde string).
    LimitQuery:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Tamaño de página (máx 100).
...
    RiskLevelQuery:
      name: riskLevel
      in: query
      schema: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
      description: Filtra por nivel de riesgo.
...
    RequestIdPath:
      name: requestId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 120 }
      description: Identificador de request (`x-request-id`) a buscar en el log de acciones.
    ActionLogEndpointIdQuery:
      name: endpointId
      in: query
      schema: { type: string, pattern: '^[1-9][0-9]*$' }
      description: Filtra por id de endpoint del catálogo.
    ActionLogRequestIdQuery:
      name: requestId
      in: query
      schema: { type: string, minLength: 1, maxLength: 120 }
      description: Filtra por requestId exacto.
    ActionLogCorrelationIdQuery:
      name: correlationId
      in: query
      schema: { type: string, minLength: 1, maxLength: 120 }
      description: Filtra por correlationId.
    ActionLogMethodQuery:
      name: method
      in: query
      schema: { type: string, enum: [GET, POST, PUT, PATCH, DELETE, OPTIONS, HEAD] }
      description: Filtra por método HTTP.
    ActionLogStatusCodeQuery:
      name: statusCode
      in: query
      schema: { type: integer, minimum: 100, maximum: 599 }
      description: Filtra por status code de la respuesta registrada.
    ActionLogActorTypeQuery:
      name: actorType
      in: query
      schema: { type: string, minLength: 1, maxLength: 80 }
      description: Filtra por tipo de actor.
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        requestId: { type: string, nullable: true, example: 'req-8f3c2a' }
        error:
          type: object
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, INTERNAL_ERROR]
            message: { type: string, example: 'SYSTEM_ENDPOINT_NOT_FOUND' }
          required: [code, message]
        timestamp: { type: string, format: date-time }
      required: [error, timestamp]
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
      required: [page, limit, total, totalPages]

    # ---------- Entidades del dominio (forma real devuelta por los mappers) ----------
...
    ActionLog:
      type: object
      description: Forma producida por `mapActionLog`.
      properties:
        actionLogId: { type: string }
        requestId: { type: string }
        correlationId: { type: string, nullable: true }
        endpointCatalogId: { type: string, nullable: true }
        actorUserId: { type: string, nullable: true }
        actorType: { type: string }
        actorRole: { type: string, nullable: true }
        method: { type: string }
        routeTemplate: { type: string, example: '/systems/endpoints/:endpointId' }
        resolvedUrlSanitized: { type: string }
        module: { type: string }
        actionName: { type: string }
        ipAddress: { type: string, nullable: true }
        targetType: { type: string, nullable: true }
        targetId: { type: string, nullable: true }
        customerId: { type: string, nullable: true }
        responseStatusCode: { type: integer }
        durationMs: { type: number }
        riskLevel: { type: string }
        containsPii: { type: boolean }
        occurredAt: { type: string, format: date-time, nullable: true }
...
    Envelope_ActionLogList:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data:
          type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/ActionLog' } }
            meta: { $ref: '#/components/schemas/PaginationMeta' }
    Envelope_ActionLogItemsOnly:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data:
          type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/ActionLog' } }
...